requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
dominate==2.3.5
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
